TV: Move AppCoordinator to its own file
