+++ Pocket Casts TV App/UI/AppCoordinator.swift
@@ -1,0 +1,18 @@
+import SwiftUI
+
+@Observable
+class AppCoordinator {
+    enum State {
+        case loading
+        case welcome
+        case browsing
+        case signedIn
+        case userSync
+    }
+
+    var state: State = .welcome
+
+    init() {
+
+    }
+}
