+++ Pocket Casts TV App/UI/RootView.swift
@@ -1,22 +1,5 @@
 import SwiftUI
 
-@Observable
-class AppCoordinator {
-    enum State {
-        case loading
-        case welcome
-        case browsing
-        case signedIn
-        case userSync
-    }
-
-    var state: State = .welcome
-
-    init() {
-
-    }
-}
-
 struct RootView: View {
     @State private var coordinator = AppCoordinator()
